ci: rewrite GitHub actions to use `nix develop|build`
Instead of relying on GitHub actions to check/test/build the code, I can
rely entirely on nix to do these steps. This has the added benefit that
the workflow will be similar to my local environment and make it easy to
test the worfklow before running it on GitHub.