ref: print succinct or detailed information about a cert
Most of the time, all I want is to know when the certificate is going to
expire. In some cases, I might want to see more details.

This change introduce a new argument to the CLI to select what level of
information to print. The new option (`--format`) will by default print
the short summary, like this:
```
> ./target/debug/x509-info twitter.com
twitter.com is valid until Mon, 12 Dec 2022 15:59:59 -0800 (29 days left)
```

While with `--format long` we get the following:
```
> ./target/debug/x509-info --format long twitter.com
certificate
 version: V3
 serial: 0d:e1:52:69:6b:2f:96:70:d6:c7:db:18:ce:1c:71:a0
 subject: C=US, ST=California, L=San Francisco, O=Twitter, Inc., CN=twitter.com
 issuer: C=US, O=DigiCert Inc, CN=DigiCert TLS RSA SHA256 2020 CA1
 validity
  not before    : Sun, 12 Dec 2021 16:00:00 -0800
  not after     : Mon, 12 Dec 2022 15:59:59 -0800
  validity days : 364
  remaining days: 29
 SANs:
  DNS:twitter.com
  DNS:www.twitter.com
```