initial commit
Setup the development environment using nix with `devshell`, and rely
on `treefmt` to format the various pieces of code.

Add some GHA to test and maintain dependencies.

Very basic implementation of the command: we accept a single argument,
a domain, check the certificates for that domain, and display the
result to STDOUT.