improve documentation
Update the README to reflect the actual flags from the program.

Add short hands options (for example `-f` in addition to `--format`).

Move the template directly in the code, it's small enough that it does
not make much sense to have it in an external file.

Add date and version to the program, and they can be displayed by
using the `-v` or `--version` flags.